Generation.EnvTable

Documentation

buildTraitTable :: [ProgramItem] -> TraitTable

buildTraitTable constructs a table of trait signatures from a list of program items.

It traverses the program items and, for every trait definition (wrapped in PI_Trait), inserts an entry mapping the trait's name to its list of abstract methods.

type TraitTable = Map String [AbsMethodDef]

TraitTable is a mapping from a trait name to its list of abstract methods.