Generation.Generation

Description

This section of the module is responsible for emitting JavaScript code for methods and translating program items (such as struct, implementation, and function definitions) into their corresponding JavaScript representations.

Documentation

generateJS :: Program -> String

generateJS: Converts a Program AST to JavaScript.

translateFuncDef :: FuncDef -> String

translateBlock :: [Stmt] -> String

translateStmt :: Stmt -> String

Translate a Stmt AST node into a string of an equivalent javascript expression

translateType :: Type -> String

translate an Type AST node into a string of an equivalent javascript expression

translateParam :: Param -> String

Translate a Param AST node into a string of an equivalent javascript expression

translateExpr :: Expr -> String

Translate an expression AST node into a string of an equivalent javascript expression

createOutputFile :: Program -> String -> IO String

Test function to see if we can write the output of a generateJS to a test javascript file

translateStructParam :: StructActualParam -> String

Translate a StructActualParam AST node into a string of an equivalent javascript expression

translateFunc :: FuncDef -> String

Translates a StructDef AST node into a string of an equivalent javascript

translateImpl :: TraitTable -> ImplDef -> String

translateImpl translates an implementation definition (ImplDef) into its

translateItem :: TraitTable -> ProgramItem -> String

translateItem translates a top-level program item (ProgramItem) into its JavaScript equivalent. Depending on the type of program item it dispatches to the appropriate translation function.

wrapBlock :: Stmt -> String

Wraps a block statement in curly braces

indent :: Int -> String -> String

Indent Helper function

translateIterForLoop :: Stmt -> String

Helper function for the assgStmt in the For Loop

type TraitTable = Map String [AbsMethodDef]

TraitTable is a mapping from a trait name to its list of abstract methods.