Parser.AST

Description

This module defines the core AST data types used in Hasgull.

Documentation

data Expr

The Expr data type represents expressions in the Hasgull language.

Constructors

Identifier String
A variable or identifier.

Int Int
An integer literal.

Trueish
Handles True

Falseish
Handles False

NewStruct Type [StructActualParam]
Creates a new struct instance.

Negative Expr
Negation of an expression.

Add Expr Expr
Addition of two expressions.

DotExpr Expr Expr
Field or method access.

Call Expr [Expr]
Function or method call.

Sub Expr Expr
Subtraction of two expressions.

LowerSelf
The special self keyword.

Multiply Expr Expr
Multiplication of two expressions.

Division Expr Expr
Division of two expressions.

Equals Expr Expr
Equality comparison.

NotEquals Expr Expr
Inequality comparison.

GreaterThan Expr Expr
Greater-than comparison.

LessThan Expr Expr
Less-than comparison.

data Type

The Type data type represents types in the Hasgull language.

Constructors

IntType
The integer type.

VoidType
The void type.

BooleanType
The boolean type.

SelfType
The special Self type.

StructName String
A named struct type.

HigherOrderType [Type] Type
A higher-order function type with parameter types and a return type.

data Param

The Param data type represents parameters (for functions or declarations) with a name and type.

Constructors

Param String Type
A parameter with a name and an associated type.

data Stmt

The Stmt data type represents statements in the Hasgull language.

Constructors

LetStmt Param Expr
Variable declaration with initialization.

AssgStmt Expr Expr
Assignment of an expression to a variable.

WhileStmt Expr Stmt
While-loop; executes a statement while the condition holds.

ForStmt Stmt Expr Stmt Stmt
For-loop with initialization, condition, update, and body.

IfStmt Expr Stmt (Maybe Stmt)
If-statement with an optional else clause.

BreakStmt
Break out of a loop.

PrintLnStmt Expr
Print an expression to the console.

BlockStmt [Stmt]
A block of statements.

ReturnStmt (Maybe Expr)
Return statement with an optional expression.

ExprStmt Expr
An expression used as a statement.

data Program

The Program data type represents an entire Hasgull program.

Constructors

Program

Fields

progItems :: [ProgramItem]

Top-level program items such as struct, trait, or function definitions.

progStmts :: [Stmt]

Top-level statements.

data ProgramItem

The ProgramItem data type represents items that can appear at the top level in a program.

Constructors

PI_Struct StructDef
A struct definition.

PI_Trait TraitDef
A trait (typeclass) definition.

PI_Impl ImplDef
An implementation of a trait.

PI_Func FuncDef
A function definition.

data TraitDef

The TraitDef data type represents a trait (or typeclass) definition.

Constructors

TraitDef

Fields

traitName :: String

The name of the trait.

traitAbsMethodDef :: [AbsMethodDef]

The abstract methods declared by the trait.

data AbsMethodDef

The AbsMethodDef data type represents an abstract method definition within a trait.

Constructors

AbsMethodDef

Fields

abMethName :: String

The name of the abstract method.

abMethParameters :: [Param]

The parameters of the method.

abMethReturnType :: Type

The return type of the method.

data StructDef

The StructDef data type represents a struct definition.

Constructors

StructDef

Fields

strucName :: String

The name of the struct.

strucFields :: [Param]

The fields of the struct.

data ImplDef

The ImplDef data type represents an implementation of a trait for a specific type.

Constructors

ImplDef

Fields

implTraitName :: String

The name of the trait being implemented.

iForType :: Type

The type for which the trait is implemented.

iMethods :: [ConcMethodDef]

The concrete method definitions that implement the trait.

data ConcMethodDef

The ConcMethodDef data type represents a concrete method definition.

Constructors

ConcMethodDef

Fields

cmName :: String

The name of the concrete method.

cmParameters :: [Param]

The parameters for the method.

cmReturnType :: Type

The return type of the method.

cmBody :: [Stmt]

The body of the method, expressed as a list of statements.

data FuncDef

The FuncDef data type represents a top-level function definition.

Constructors

FuncDef

Fields

funcName :: String

The function name.

funcParameters :: [Param]

The parameters of the function.

funcReturnType :: Type

The return type of the function.

funcBody :: [Stmt]

The function body as a list of statements.

data StructActualParam

The StructActualParam data type represents an actual parameter used when creating a struct instance.

Constructors

StructActualParam String Expr
A field initialization with a field name and an expression.